test: hoist expectSameCurve to a shared fixture across Curve2D interpolation parity suites (#1256)
